feat: refactor layout
