+++ packages/iceworks-client/.eslintrc
@@ -16,9 +16,11 @@
   "plugins": ["react", "babel"],
   "rules": {
     "react/jsx-filename-extension": 0,
-    "jsx-a11y/anchor-is-valid": 0,
-    "arrow-body-style": 0,
     "react/no-array-index-key": 0,
+    "jsx-a11y/anchor-is-valid": 0,
+    "jsx-a11y/click-events-have-key-events": 0,
+    "jsx-a11y/no-static-element-interactions": 0,
+    "arrow-body-style": 0,
     "import/prefer-default-export": 0
   }
 }
