+++ packages/iceworks-client/public/index.html
@@ -4,10 +4,11 @@
     <meta charset="utf-8" />
     <meta http-equiv="x-ua-compatible" content="ie=edge,chrome=1" />
     <meta name="viewport" content="width=device-width" />
+    <link href="//at.alicdn.com/t/font_568722_zekhvu8mmzg.css" rel="stylesheet" />
     <title>Iceworks</title>
   </head>
 
   <body>
-    <div id="ice-container"></div>
+    <div id="iceworks"></div>
   </body>
 </html>
