+++ packages/iceworks-client/src/components/Aside/index.module.scss
@@ -1,0 +1,5 @@
+.aside {
+  background: $color-brand1-6;
+  width: 80px;
+  box-shadow: 2px 0 6px rgba(0, 21, 41, 0.35);
+}
