+++ packages/iceworks-client/src/components/ConnectionBar/index.module.scss
@@ -1,6 +1,9 @@
 .connectionBar {
+  flex: auto 0 0;
+  height: 40px;
   background: #e83030;
   color: #fff;
-  padding: 10px 0;
-  text-align: center;
+  display: flex;
+  align-items: center;
+  justify-content: center;
 }
