+++ packages/iceworks-client/src/components/Icon/index.js
@@ -6,12 +6,10 @@ const fontFamily = 'iceworks';
 const prefix = 'iceworks';
 
 // icon url: http://www.iconfont.cn/manage/index?manage_type=myprojects&projectId=568722&keyword=
-const Icon = ({
-  className, size, type, ...other
-}) => {
+const Icon = ({ className, size, type, ...other }) => {
   const classes = cx({
-    [`${prefix}-${size}`]: !!size,
     [`${fontFamily}`]: true,
+    [`${prefix}-${size}`]: !!size,
     [`${prefix}-${type}`]: !!type,
     [className]: !!className,
   });
