+++ packages/iceworks-client/src/components/ThemeToggler/index.module.scss
@@ -1,0 +1,5 @@
+.themeToggler {
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+}
