+++ packages/iceworks-client/src/global.scss
@@ -1,0 +1,5 @@
+html,
+body,
+#iceworks {
+  height: 100%;
+}
