+++ packages/iceworks-client/src/pages/Dev/index.js
@@ -26,6 +26,6 @@ export default class Dev extends Component {
   }
 
   onDev() {
-    
+
   }
 }
